@@ -8,6 +8,7 @@
     "start:commercial": "ng serve web-commercial",
     "start:prod": "ng build --prod --sourceMap=true && http-server ./dist/apps/web/web-portal -a www.skysmack.test -p 4000 -o",
     "build": "gulp webLocalization && ng build --prod --deploy-url //cdn.skysmack.net/",
+    "build:commercial": "gulp webLocalization && ng build --prod --deploy-url //skysmack.com/",
     "build:dev": "ng build --stats-json && npm run local-dev-server-insecure",
     "build:prod": "ng build --prod --stats-json && npm run local-dev-server-insecure",
     "build:analyze": " webpack-bundle-analyzer ./dist/apps/web/web-portal/stats.json",
@@ -62,26 +63,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.1.2",
-    "@angular/common": "8.1.2",
-    "@angular/compiler": "8.1.2",
-    "@angular/core": "8.1.2",
-    "@angular/forms": "8.1.2",
-    "@angular/platform-browser": "8.1.2",
-    "@angular/platform-browser-dynamic": "8.1.2",
-    "@angular/router": "8.1.2",
+    "@angular/animations": "8.2.0",
+    "@angular/common": "8.2.0",
+    "@angular/compiler": "8.2.0",
+    "@angular/core": "8.2.0",
+    "@angular/forms": "8.2.0",
+    "@angular/platform-browser": "8.2.0",
+    "@angular/platform-browser-dynamic": "8.2.0",
+    "@angular/router": "8.2.0",
     "core-js": "2.6.8",
     "rxjs": "6.5.2",
-    "zone.js": "0.9.1",
+    "zone.js": "0.10.1",
     "@angular-redux/router": "10.0.0",
     "@angular-redux/store": "10.0.0",
-    "@angular/cdk": "8.1.1",
-    "@angular/material": "8.1.1",
-    "@angular/material-moment-adapter": "8.1.1",
-    "@angular/pwa": "0.801.2",
-    "@angular/service-worker": "8.1.2",
+    "@angular/cdk": "8.1.2",
+    "@angular/material": "8.1.2",
+    "@angular/material-moment-adapter": "8.1.2",
+    "@angular/pwa": "0.802.0",
+    "@angular/service-worker": "8.2.0",
     "@aspnet/signalr": "1.1.4",
-    "@kolkov/angular-editor": "0.17.1",
+    "@kolkov/angular-editor": "0.18.4",
     "@ngx-translate/core": "11.0.1",
     "@redux-offline/redux-offline": "2.5.2",
     "angular-calendar": "0.27.14",
@@ -95,28 +96,29 @@
     "parse-link-header": "1.0.1",
     "redux": "4.0.1",
     "redux-observable": "1.1.0",
-    "swagger-ui": "3.23.2"
+    "swagger-ui": "3.23.4"
   },
   "devDependencies": {
     "jest-preset-angular": "7.1.1",
-    "@nrwl/cypress": "8.2.0",
-    "@angular/compiler-cli": "8.1.3",
-    "@angular/language-service": "8.1.2",
-    "@angular-devkit/build-angular": "0.801.2",
+    "@nrwl/cypress": "8.4.4",
+    "@angular/compiler-cli": "8.2.0",
+    "gulp-cli": "2.2.0",
+    "@angular/language-service": "8.2.0",
+    "@angular-devkit/build-angular": "0.802.0",
+    "@angular-devkit/build-ng-packagr": "0.802.0",
     "codelyzer": "5.1.0",
-    "@nrwl/schematics": "8.4.0",
+    "@nrwl/schematics": "8.4.4",
     "@nrwl/builders": "8.0.2",
     "jest": "24.8.0",
-    "@types/jest": "24.0.15",
+    "@types/jest": "24.0.17",
     "ts-jest": "24.0.2",
-    "@angular-devkit/build-ng-packagr": "0.801.2",
-    "@angular/cli": "8.1.3",
-    "@types/node": "12.6.8",
+    "@angular/cli": "8.2.0",
+    "@types/node": "12.7.0",
     "cypress": "3.4.1",
     "dotenv": "8.0.0",
     "gulp-merge-json": "1.3.1",
     "gulp-plumber": "1.2.1",
-    "ng-packagr": "5.3.0",
+    "ng-packagr": "5.4.3",
     "normalize-scss": "7.0.1",
     "prettier": "1.18.2",
     "ts-node": "8.3.0",
@@ -124,6 +126,6 @@
     "tslib": "1.10.0",
     "tslint": "5.18.0",
     "typescript": "3.4.5",
-    "webpack-bundle-analyzer": "3.3.2"
+    "webpack-bundle-analyzer": "3.4.1"
   }
 }
