@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
-import { Oauth2Module } from './../../../../../../../libs/portal-packages/src/lib/oauth2/oauth2.module';
+import { OAuth2Module } from './../../../../../../../libs/portal-packages/src/lib/oauth2/oauth2.module';
 
 @NgModule({
   imports: [
-    Oauth2Module
+    OAuth2Module
   ]
 })
 export class OAuth2WrapperModule { }
