@@ -1,3 +1,6 @@
-export * from './commercial-account-index/commercial-account-index.component';
 export * from './commercial-account-components';
 
+export * from './commercial-account-dashboard/commercial-account-dashboard.component';
+export * from './commercial-account-change-password/commercial-account-change-password.component';
+export * from './commercial-account-forgot-password/commercial-account-forgot-password.component';
+export * from './commercial-account-login/commercial-account-login.component';
