@@ -6,4 +6,5 @@ export * from './notifications/index';
 export * from './helpers/index';
 export * from './packages/index';
 export * from './signal-r/index';
-export * from './dashboard';
+export * from './dashboard/index';
+export * from './redux/index';
