@@ -2,5 +2,4 @@ export * from './applications-components';
 export * from './applications-index/applications-index.component';
 export * from './applications-edit/applications-edit.component';
 export * from './applications-create/applications-create.component';
-export * from './applications-select/applications-select.component';
-export * from './applications-select-field/applications-select-field.component';
+export * from './applications-roles/applications-roles.component';
