@@ -1,4 +1,4 @@
-import { Component, OnInit, OnDestroy, Injector } from '@angular/core';
+import { Component, OnInit, Injector } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { LocalObject, RouteData, toLocalObject } from '@skysmack/framework';
 import { NgSkysmackStore } from '@skysmack/ng-skysmack';
@@ -15,7 +15,7 @@ import { FieldsConfig } from '@skysmack/ng-fields';
   selector: 'ss-settings',
   templateUrl: './settings.component.html'
 })
-export class SettingsComponent extends BaseComponent<SettingsAppState<any>, unknown> implements OnInit, OnDestroy {
+export class SettingsComponent extends BaseComponent<SettingsAppState<any>, unknown> implements OnInit {
 
   public editorItem: LocalObject<any, unknown>;
   public selectedSettings: LocalObject<any, unknown>;
@@ -41,10 +41,6 @@ export class SettingsComponent extends BaseComponent<SettingsAppState<any>, unkn
     this.settingsKey = settingsKey ? settingsKey : 'default';
     this.actions.get(this.packagePath, this.settingsKey);
     this.setFields();
-  }
-
-  ngOnDestroy() {
-    super.ngOnDestroy();
   }
 
   public onSettingsSubmit(fh: FormHelper) {
