@@ -6,4 +6,5 @@ export * from './shared-reducer';
 export * from './queue-reducer';
 export * from './field-reducer';
 export * from './cancel-field-action';
-export * from './settings-reducer'
+export * from './settings-reducer';
+export * from './hydrated-reducer';
